third_party/WebKit/LayoutTests/http/tests/websocket/permessage-deflate-invalid-parameter-expected.txt

CONSOLE ERROR: line 28: WebSocket connection to 'ws://localhost:8880/permessage-deflate-invalid-parameter?x-foo' failed: Error during WebSocket handshake: Error in permessage-deflate: Received an unexpected permessage-deflate extension parameter
CONSOLE ERROR: line 28: WebSocket connection to 'ws://localhost:8880/permessage-deflate-invalid-parameter?client_max_window_bits=7' failed: Error during WebSocket handshake: Error in permessage-deflate: Received invalid client_max_window_bits parameter
CONSOLE ERROR: line 28: WebSocket connection to 'ws://localhost:8880/permessage-deflate-invalid-parameter?client_max_window_bits=16' failed: Error during WebSocket handshake: Error in permessage-deflate: Received invalid client_max_window_bits parameter
CONSOLE ERROR: line 28: WebSocket connection to 'ws://localhost:8880/permessage-deflate-invalid-parameter?client_no_context_takeover=foo' failed: Error during WebSocket handshake: Error in permessage-deflate: Received invalid client_no_context_takeover parameter
CONSOLE ERROR: line 28: WebSocket connection to 'ws://localhost:8880/permessage-deflate-invalid-parameter?client_max_window_bits=8;%20client_no_context_takeover;%20x-foo' failed: Error during WebSocket handshake: Error in permessage-deflate: Received an unexpected permessage-deflate extension parameter
Test whether WebSocket rejects invalid permessage-deflate parameters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing parameter: "x-foo"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "client_max_window_bits=7"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "client_max_window_bits=16"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "client_no_context_takeover=foo"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Testing parameter: "client_max_window_bits=8; client_no_context_takeover; x-foo"
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
